Allows 64 values for inclusive criterion
Current implementation allows only 32 values for an inclusive
criterion. It comes from the use of a 32 bit integer for the
internal state representation.
This patch replaces the state representation by a 64 bit integer
which allows 64 bit. This is a temporary patch as the entire
interface should be reworked to avoid such problems in the
future.

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-31110

+++ parameter/SelectionCriterion.h
@@ -45,8 +45,8 @@ public:
 
     /// From ISelectionCriterionInterface
     // State
-    virtual void setCriterionState(int iState);
-    virtual int getCriterionState() const;
+    virtual void setCriterionState(uint64_t iState);
+    virtual uint64_t getCriterionState() const;
     // Name
     virtual std::string getCriterionName() const;
     // Type
@@ -56,10 +56,10 @@ public:
     void resetModifiedStatus();
 
     /// Match methods
-    bool is(int iState) const;
-    bool isNot(int iState) const;
-    bool includes(int iState) const;
-    bool excludes(int iState) const;
+    bool is(uint64_t iState) const;
+    bool isNot(uint64_t iState) const;
+    bool includes(uint64_t iState) const;
+    bool excludes(uint64_t iState) const;
 
     /// User request
     std::string getFormattedDescription(bool bWithTypeInfo, bool bHumanReadable) const;
@@ -77,7 +77,7 @@ public:
     virtual void toXml(CXmlElement& xmlElement, CXmlSerializingContext& serializingContext) const;
 private:
     // Current state
-    int _iState;
+    uint64_t _iState;
     // Type
     const CSelectionCriterionType* _pType;
